requirements.txt
selenium
pytesseract
pillow
opencv-python
flask